fix(reorder-buttons): stricten preventDefault use
Configured hotkey keys should now be allowed through on use.
Previously, the `preventDefault` event call activated even if the
modifiers for a hotkey weren't pressed. For hotkeys such as
Atl-Shift-Minus, this prevented typing regular '-' characters for
example. This has been updated so `preventDefault` is only called
when the entire hotkey is actually executed.