chore(reorder-buttons): remove user filename suffix
This change is to enable loading the script file in a @require statement
so that the script can be edited in a local editor, saved to the
file-system and then loaded by TamperMonkey within a skeleton userscript
that just loads and runs the local file, making iteration _way_ quicker.

Towards this setup, we are also including a sample local test-script header for use when configuring TamperMonkey.

Details: https://danieldusek.com/use-external-editor-to-write-user-scripts-tampermonkey.html