feat(reorder-buttons): add one-click sort button
Back by popular demand (kinda): one-click button to sort reviews. Based off
questions after the transitioning the review sort buttons into the cogwheel
menu, it seems there's a decent number of users that expected some immediate
button to use for sorting.

The confusion is reasonable enough given the script title would indicate that
the main thing being done is adding buttons to control SRS sorting, so we're
adding back in a button to do just that. It'll display by default, though a
setting has been added for users that would rather have it hidden. The button
also acts as an indicator of the current sorting mode that is active _and_
allows for shuffling the review queue back to a random order (including
disabling of 1x1 mode).

The previous sorting buttons were custom HTML elements with hard-coded styling
that integrated poorly into the existing UI. Hopefully this better-integrated,
text-less replacement is sufficient.