feat(reorder-buttons): enable custom startup sort
It is now possible to configure descending sort on startup. This
required a settings change that isn't necessarily backwards compatible.
This is a small breaking change that requires only a small level of
effort to address on user's end, so it will be included as part of a
larger set of changes on release.

Documentation has been updated to reflect this new feature.