chore(reorder-buttons): force unique HTML elem ids
Pre-empting some bug report here caused by incompatibility with another
script using ids to modify HTML elements that are in conflict with the
ids used by this script when modifying its own elements.

This solutions doesn't guarantee anything, but with such a long and
specific id string prefix, it should be fine.