chore(reorder-buttons): add userscript extension
So it seems that TamperMonkey is still able to load the userscript even
if it has the `.user.js` extension from a `require` directive pointing
to a local file. With that sorted, it'd be better to continue using
the `script.user.js` filename as it'll permit installing the script
with one-click on the browser without a third-party site or some CI to
serve the file somewhere.

The README has been re-updated to reflect this latest change.