feat(reorder-buttons): enable question type sort
Users now have the ability for force certain question types to come
first when reviewing items. This will force question type ordering such
that the same question type will always be asked first and on-repeat
until a correct answer for that item and question type is submitted. For
example, if question type ordering is set to reading-first, then any
time a given item comes up for review, the question asked will be for
the reading of the item until a correct reading answer is given for that
item.

This behavior can be enabled or disabled mid-review and is completely
independent of SRS sorting or 1x1 mode.