fix(reorder-buttons): update 1x1 on settings save
Previously, 1x1 mode was activated or deactivate whenever the review
queue was sorted. This meant that it's activation and use depended on
sorting review items, which means that any user that'd want to use 1x1
mode regardless of SRS sorting was forced to use both simultaneously.

This fix makes it so that 1x1 mode is enabled or disabled whenever
settings are updated as well as on startup.