fix(reorder-buttons): ensure question type on sort
There's been a recent bug report where radical questions have their
answer box set to accept inputs for a reading question. Upon inputting
an answer, the review page crashes.

Seems like upon sorting, the `order_question_type` func is called,
though by the time it is called, the `questionType` field in jStorage is
out of date and we incorrectly assume that because the item is a radical
question, that the `questionType` field will be accurate.

The function's been updated to ensure that it'll set the `questionType`
field to "meaning" for radical items in case it is out of sync.

This commit also removes a premature call to the `order_question_type`
function during registration. I've no clue why that call we made
regardless of user settings.