fix(reorder-buttons): wrap question type ordering
Thanks to some extra logging, I found a critical recursive error: the
`order_question_type` function simultaneously registered as a listener
and triggered events on the `currentItem` key in jStorage. This caused a
huge loop that would eventually fill the entire stack for the userscript
thread. Beyond that, there'd be horrendous latency while the async
function finished killing the process.

By wrapping the function in another function unregisters the listener
before calling the order func and then re-registering afterwards, we
avoid the recursion.