fix(reorder-buttons): add back2back compatibility
The Back2Back userscript by Kumirei wrap the default `jStorage.set`
method to implement reordering functionality. To enable compatibility
with Kumirei's userscript, we've agreed to use a new option in the `set`
method to let Back2Back know it should avoid reordering question types
on a per-call basis of `set`.