feat(reorder-buttons): remove 1x1 mode support
While technically a removal of a feature, the real value here is
reducing the overhead of maintaining a critically broken and
undocumented function and having users understand what 1x1 mode even
means. All trace of the setting and functionality should be removed from
the userscript entirely in this commit -- a separate commit is dedicated
to updating documentation.